Features/add services for post management

+++ src/main/java/kr/co/mz/sns/dto/PostDto.java
@@ -5,12 +5,10 @@ import lombok.Data;
 
 @Data
 public class PostDto {
-    private int seq;
+    private Long seq;
     private String content;
     private Integer likes;
     private Timestamp createdAt;
     private Timestamp modifiedAt;
-
     private int userSeq;
-
 }
